Simplify macOS write to status=progress, drop SIGINFO machinery
macOS dd has supported status=progress since Monterey (12.0). Switching
to it lets us delete the entire SIGINFO/process-tree-walk apparatus that
was needed to drive BSD dd's per-second status output through sudo's
PAM monitor.

The Linux and macOS parsers were already converging — both dd variants
emit a status line whose first whitespace-delimited token is the byte
count, separated by '\r' until the final '\n'-terminated summary. Move
the shared parser to dd_progress.go (build: darwin || linux) and reuse
it from both platforms.

Also tighten the cmd.Wait / scanner-goroutine ordering on both
platforms: wait for the scanner to finish reading before returning, so
no progress callbacks fire after writeImageToDisk has returned.

Net: -336 lines (incl. ~140 lines of tests for code that no longer
exists), three files instead of two with duplicated logic.
